Split App root and PanelManager
